move deduplicated finding report list to own component
