+++ src/components/sections/judge/JudgeDeduplicatedFinding.tsx
@@ -1,12 +1,13 @@
 'use client'
 
 import Link from 'next/link'
-import {ArrowLeft, ArrowRight, Trash2, VerifiedIcon} from 'lucide-react'
+import {ArrowLeft} from 'lucide-react'
 import {notFound} from 'next/navigation'
 
 import JudgeAddDeduplicatedFinding from './JudgeAddDeduplicatedFinding'
 import FindingSeverityBadge from '../finding/FindingSeverityBadge'
 import JudgeEditDeduplicatedFindingFormDialog from './JudgeEditDeduplicatedFindingFormDialog'
+import JudgeDeduplicatedFindingReportsList from './JudgeDeduplicatedFindingReportsList'
 
 import {Button} from '@/components/ui/Button'
 import {PATHS} from '@/lib/utils/common/paths'
@@ -17,8 +18,6 @@ import {useGetFindings} from '@/lib/queries/finding/getFinding'
 import {Tabs, TabsContent, TabsList, TabsTrigger} from '@/components/ui/Tabs'
 import {useSearchParamsEnumState} from '@/lib/hooks/useSearchParamsState'
 import {JudgeDeduplicatedFindingTab} from '@/server/db/models'
-import {useRemoveDededuplicatedFinding} from '@/lib/queries/deduplicatedFinding/removeDeduplicatedFinding'
-import {toast} from '@/components/ui/Toast'
 
 type JudgeDeduplicatedFindingProps = {
   deduplicatedFindingId: string
@@ -37,24 +36,6 @@ const JudgeDeduplicatedFinding = ({
   const {data: findings, isLoading: findingsLoading} = useGetFindings({
     deduplicatedFindingId,
   })
-
-  const {mutate: removeDeduplicatedFindingMutate} =
-    useRemoveDededuplicatedFinding()
-
-  const removeDeduplicatedFinding = (findingId: string) => {
-    removeDeduplicatedFindingMutate(
-      {findingId},
-      {
-        onSuccess: () => {
-          toast({
-            title: 'Success',
-            description:
-              'Finding has been removed from this deduplicated finding.',
-          })
-        },
-      },
-    )
-  }
 
   if (deduplicatedFindingLoading || findingsLoading) {
     return (
@@ -130,47 +111,11 @@ const JudgeDeduplicatedFinding = ({
               value={JudgeDeduplicatedFindingTab.REPORTS}
               className="flex flex-col gap-4">
               {findings.map((finding) => (
-                <div
+                <JudgeDeduplicatedFindingReportsList
                   key={finding.id}
-                  className="flex h-[72px] w-full items-center justify-between bg-grey-90 p-4">
-                  <Button asChild variant="link">
-                    <Link
-                      href={PATHS.judgeFinding(finding.id)}
-                      className="flex gap-6">
-                      {deduplicatedFinding.bestFindingId === finding.id && (
-                        <div className="flex items-center gap-1">
-                          <VerifiedIcon width={24} height={24} />
-                          <span className="text-bodyS">Best</span>
-                        </div>
-                      )}
-                      <span className="text-titleS">{finding.title}</span>
-                    </Link>
-                  </Button>
-                  <div className="flex items-center gap-6">
-                    <div>
-                      <span className="text-bodyM opacity-70">
-                        Submited by:{' '}
-                      </span>
-                      <span className="text-bodyM">
-                        {finding.author.alias ?? finding.author.id}
-                      </span>
-                    </div>
-                    <Button
-                      className="flex items-center gap-2"
-                      size="small"
-                      variant="outline"
-                      onClick={() => removeDeduplicatedFinding(finding.id)}>
-                      <span>Remove</span>
-                      <Trash2 width={16} height={16} />
-                    </Button>
-                    <Button asChild size="small" variant="default">
-                      <Link href={PATHS.judgeFinding(finding.id)}>
-                        <span>Detail</span>
-                        <ArrowRight width={16} height={16} />
-                      </Link>
-                    </Button>
-                  </div>
-                </div>
+                  finding={finding}
+                  deduplicatedFinding={deduplicatedFinding}
+                />
               ))}
             </TabsContent>
             <TabsContent value={JudgeDeduplicatedFindingTab.ADD_FINDINGS}>
